Avoid warning during PME tuning
During PME tuning a warning could be printed about the PME grid not
being a multiple of the number of PME ranks. Since this is not
actually an issue and the user can not influence this, this note is
removed during tuning. To achieve this, gmx_pme_init() now actually
using the logging framework.