Avoid using AUTO settings in CI builds
AUTO settings e.g. for SIMD will change the build
configuration depending on the underlying hardware
the container happens to run on, which makes the
builds non-reproducible in case of problems.

This changes the two AUTO builds to use AVX2
instead, which should realistically be present on
any modern hardware where gitlab CI tests run.