Replace automatic rdtscp checks with boolean option
This simplifies the CMake configuration by skipping the
automatic detection based on build host. Virtually all
64-bit x86 machines support this anyway, apart from the
very first AMD 64-bit release that are now 7 years old,
so it should be fine to enable by default. In the rare
case of 7-year-old hardware, gmx will detect missing
support at the start of the run.