Allow simpler use of scripted container builds
There's no need to have a list of supported versions of things in the
script that supports building CI images. We should have a list, e.g.
in the install guide, but not here.

One of the primary use cases for the script is building new
images. Those are most often needed for new versions of things, and
expanding the supported versions list is needless work.

Removed use of const and relaxed types for versions that e.g. need
not be integers.

Made suggestions in the doc strings of valid choices.

Note, had to force-push this branch to get out of non-linear merge
history hell.

Fixes #3622