Revert "Modernize STL usage" for fixedCapacityVector
This reverts 2f876de263efdf08daf160e7791434cd760ccd9b for
fixedCapacityVector.h only. The syntax that was added is not yet
supported by the CUDA compiler (including the latest 11.0 version),
and as such this revertion is required to allow use of
fixedCapacityVector from within CUDA files (required for an upcoming
change). This should be added back when CUDA support for the syntax
arrives.