Fix bug in Lincs GPU setup
Memory re-allocation was previously only triggered if the number of
constraints threads increases. But the sizes also depend on the value
of "maxCoupledConstraints", and its possible that this can increase
without the number of constraints threads increasing. This change
fixes by also triggering the reallocation when the value of
maxCoupledConstraints is increased.

(cherry picked from 87372d0ffd610943928aeb6b56fc3fb96b0c30f5)