Implement Nose-Hoover thermostat for leap-frog modular simulator
This implements the Nose-Hoover temperature coupling algorithm for the
leap-frog version of modular simulator. Velocity verlet could in theory
use this coupling algorithm as well, but the legacy version uses the
Trotter-decomposed version instead, which will be introduced to modular
simulator in a follow-up commit. To ensure equivalence of the legacy and
the modular code paths, Nose-Hoover will not be enabled for md-vv in
modular simulator until the Trotter scheme is merged.

The implementation added in this change is straightforward, as it extends
the existing general velocity-scaling temperature coupling object by
another algorithm. It is marginally more complex than the existing
algorithms (Berendsen and v-scale) due to the need of checkpointing the
extended ensemble degree of freedom for exact continuation.

Refs #3423