Tell gmxapi to look for MPI, when appropriate.
The recommended CMake configuration for MPI-enabled GROMACS does not
provide MPI compiler wrappers for regular compiler variables. However,
the MPI toolchain details do not get conveyed transitively through the
libgromacs target. This should be easily resolved by checking whether
libgromacs is built with MPI library support and, if so, using the
CMake FindMPI module.

This may not be robust and should be revisited in the context of issue
#3672.

Meanwhile, this change should reduce errors for users following the
documentation updated under #3671.