Narrow-down Intel-specific #ifdef's
Some of the things hidden by `#ifdef __INTEL_COMPILER` does not seem to be relevant as of 2021.1-beta09/10 (ICC 2021.1 Beta 20200827).

Also, was getting a false-positive "warning: result of comparison of 0-bit unsigned value != 0 is always false" for line `if (grid_base[g_baseNR - 1] % 4 != 0)`. Replaced it with `static_assert`, because why not.