Establish `api/` as the home for installed headers.
Prepare to remove `src/` from global build tree include path.
* Create `api/legacy` for the new `legacy_api` target.
* Move gmxapi and nblib public header dependencies to `api/legacy`.

This change is orthogonal to the clarification of the public API.
The purpose is to normalize the way we manage installed headers.
gmxapi and nblib already provide their public interfaces through files
and CMake targets in `api/`. For the remaining installed libgromacs
headers, the `legacy_api` target provides a transitional home as the
modules are individually updated in terms of CMake machinery and as the
legacy API functionality is reexamined.

Refs #3288