Use dynamically allocated unique pointers for GPU halo tests
The tests were intermittently failing in the enqueuing of a remote
GPU's event to a local GPU's stream because the remote object
(allocated on the stack in the innermost loop) was already destroyed (due
to the asynchronicity of the operation). This change allows the objects
to persist, and is the same way we already do it in the main code.

Fixes intermittent failures appearing in !773 and eleswhere.