Fix mdrun-only build
Recent changes to include path setup didn't handle the mdrun-only
build. The name clash between the new src/gromacs/mdrun interface
library and the former src/programs/mdrun executable is fixed. When
doing a shared-library mdrun-only build, the dependencies of
libgromacs also need to be limited in the same way as the set of
modules added to libgromacs.