Relax simulator comparison tests
The simulator comparison tests have proven to be too strict when running
them on more diverse hardware / compiler / parallelization combinations.

This doubles the tolerances for the energies of tests containing
constraints. Note that at 160 double precision ULPs, these are still
very tight for end-to-end tests, so the danger of relaxing them a bit
is low.

Closes #3836