Clarify checkpoint mismatch error, and enable in release
As of GROMACS 2021, checkpoint files written with the modular and the
legacy codepath need to be continued with the same code path. The
respective check was not executed in release mode, which would lead
to less descriptive errors.

Also reworded the error to guide users how they can solve the problem.