Fix and activate tests for kv-tree checkpoint functions
The tests in src/gromacs/fileio/tests/checkpoint.cpp were introduced
in 64bee06aa, but the file was never added to the CMakeLists.txt,
meaning they were never executed (or even compiled).

Added the file to the CMake target and fixed a few compilation
errors and warning to activate the tests.