Update deprecation statuses
Various functionality has been agreed as deprecated at #3818 and
elsewhere. Those are now noted in the release notes, documentation,
and CMake status messages. The changes are all minor so there is
little need to make mdrun report when such configurations are used.

The deprecation content for release  2020 (and 2021) series is updated
to include  the relevant  content from  release 2019  series. Comments
have been added to help us to this better in future.

Fixes #3818