Simplify test mdp handling
In the current implementation, mdp options are handled in 3 places:
* Defaults for a system are stored in the system database
* General defaults are added to that (using `insert` to avoid overwriting)
  in `prepareDefaultMdpFieldValues`.
* Some more defaults are hardcoded in `prepareMdpFileContents`

`prepareDefaultMdpFieldValues` returns a map of options and option
values.  This allows specific tests to change default options as
needed, but not to add any new options, because
`prepareDefaultMdpFieldValues` only reads a predetermined set of
options. Newly added options are dropped without warning or error. A
specific "other" field in the map allows for additional options - but
users must be careful not to add any option that is already in the map
or that are hardcoded in `prepareMdpFileContents`, which would result
in a grompp error due to having multiple values for the same option.

This makes the functionality hard to use and is prone to errors (case
in point: the periodic actions test is trying to set nstlog, which is
silently ignored, so the test coverage is not what was intended).
Changes to this functionality have previously been stalled in favor of
transition to IMdpOptions, but given that the current road map doesn't
make a full transition in the near future likely, it's likely worth a
small investment to make the current functionality more usable.

This moves all hardcoded values from `prepareMdpFileContents` into the
map returned by `prepareDefaultMdpFieldValues`, and prints all options
present in the map to the mdp string. This allows tests to change or
add any value. The users of the functionality are responsible to fill
the map with valid mdp options - and will get grompp errors if they
fail. This also removes all uses of the "other" keyword (which would
now generate a grompp error, so we're sure we didn't forget any uses).