densityfit: Fix typo in use of iterators
... and some typos in comments.

While both referenceDensity_ and referenceDensity are views into the same data,
it's clearer to use the same variable for both begin() and end().

This is a remake of MR !740. That MR was reverted (in !748) due to
failed tests. The test failures, however, were caused by an unrelated bug
(issue #3791, fix !827). So, this change is brought back.