Remove mdrun-only build configuration
The existence of this configuration is unnecessary, provides no real
advantages, and was only tested by compiling it (ie not by running
it). Its existence complicates installation, maintenance, testing,
documentation, teaching, and usage. It was deprecated in GROMACS 2021,
and is now removed.

Added a new regressiontests gitlab job now that the build
configuration that used to build with GMX_BUILD_MDRUN_ONLY can run
those tests.

Refs #3808